requirements.txt
paho_mqtt
Pillow
qrcode
flask
